local weight learning updated
